Move header icons to header :D
